Add initial test for muster controller. Ensure observation table is wiped before tests run. Move seed logic into test-utils, which seed-dev.ts now calls on. Change seed-dev script to not drop database as this is handled by stop-db command now. Remove silent flag on ts-mocha call.
