+++ packages/server/seed-dev.sh
@@ -10,22 +10,11 @@ export SQL_PASSWORD=${SQL_PASSWORD:=postgres}
 export SQL_DATABASE=${SQL_DATABASE:=dds}
 export PGPASSWORD=$SQL_PASSWORD
 
-read -r -p "Are you sure? This will delete existing data in your local '$SQL_DATABASE' database. [y/N] " response
-case "$response" in
-  [yY][eE][sS] | [yY])
-    if [[ "$(database_exists)" ]]; then
-      drop_database
-    fi
 
-    if ! [[ "$(database_exists)" ]]; then
-      create_database
-    fi
+if ! [[ "$(database_exists)" ]]; then
+  create_database
+fi
 
-    ./migration-run.sh
+./migration-run.sh
 
-    ./node.sh ./src/sqldb/seed-dev.ts
-    ;;
-  *)
-    echo "Aborted"
-    ;;
-esac
+./node.sh ./src/sqldb/seed-dev.ts
